Creacion de metodos: calcular total en clase factura y generarDetalles

+++ AppFacturas/src/main/java/org/cursodesarrollo/appFacturas/modelo/Producto.java
@@ -3,13 +3,12 @@ package org.cursodesarrollo.appFacturas.modelo;
 public class Producto {
     private int codigo;
     private String nombre;
-    private Double precio;
+    private float precio;
     private static int ultimoCodigo; //para auto incrementar el codigo
 
     public Producto(int codigo) {
         this.codigo = ++ultimoCodigo;
     }
-
 
     public int getCodigo() { //El codigo solo se podra leer, ya que se creara automaticamente.
         return codigo;
@@ -23,11 +22,13 @@ public class Producto {
         this.nombre = nombre;
     }
 
-    public Double getPrecio() {
+    public float getPrecio() {
         return precio;
     }
 
-    public void setPrecio(Double precio) {
+    public void setPrecio(float precio) {
         this.precio = precio;
     }
+
+
 }
